refactored treeview example into multiple classes
