+++ treeviews/src/main/java/no/tornado/fxsamples/treeviews/Styles.kt
@@ -1,0 +1,19 @@
+package no.tornado.fxsamples.treeviews
+
+import tornadofx.Stylesheet
+import tornadofx.box
+import tornadofx.cssclass
+import tornadofx.px
+
+class Styles : Stylesheet() {
+    companion object {
+        val wrapper by cssclass()
+    }
+
+    init {
+        s(wrapper) {
+            padding = box(10.px)
+            spacing = 10.px
+        }
+    }
+}
