+++ treeviews/src/main/java/no/tornado/fxsamples/treeviews/WorkBench.kt
@@ -1,0 +1,13 @@
+package no.tornado.fxsamples.treeviews
+
+import tornadofx.App
+import tornadofx.importStylesheet
+
+class Workbench : App() {
+    override val primaryView = Treeview::class
+
+    init {
+        importStylesheet(Styles::class)
+    }
+
+}
